Testing dummy input for double returnedStack output
Adding a stick to the fancy-bed-old-drapes-opened recipe to hopefully allow for 3 separate outputs of this recipe. If this works, we'll use this method to fix all the triple output recipes. 

If it doesn't work, let's check the log for errors. Failing that, let's just change these bed recipes to give extra colored cloth instead.